[fix] Fix extension logic and refactor the nodes(OSINT-311)
