+++ server/db_conn/neo4j/models/__init__.py
@@ -1,6 +1,4 @@
-from .user import Person, SurfaceUser,DarkUser, Company
-from .domain import Domain
-from .post import Post, Comment
-from .user_info import Email,Wallet, Phone, Message
+from .node import *
+from .relationship import *
 
-__all__ = ['Person', 'SurfaceUser', 'DarkUser', 'Company', 'Domain', 'Post', 'Comment', 'Email', 'Wallet', 'Phone', 'Message']
+__all__ = ['Person', 'SurfaceUser', 'DarkUser', 'Company', 'Domain', 'Post', 'Comment', 'Email', 'Wallet', 'Phone', 'Message', 'Relationship','NODE_LIST','RELATIONS']
