+++ server/db_conn/neo4j/models/lib/func.py
@@ -1,4 +1,4 @@
-from ..init import db
+from ...init import db
 
 def delete_node(node_id):
     query = """
