+++ server/db_conn/neo4j/models/lib/node_manager.py
@@ -7,13 +7,13 @@ class NodeManager:
         self.cls = cls
 
     def check_node(self, data):
-        node = self.cls.nodes.get_or_none(**data)
+        node = self.cls.nodes.first_or_none(**data)
         if node is not None:
             return True, node
         return False, None
 
     def get_node(self, data):
-        node = self.cls.nodes.get_or_none(**data)    
+        node = self.cls.nodes.first_or_none(**data)    
         if node:
             return node
         else:
@@ -24,7 +24,7 @@ class NodeManager:
         return nodes
     
     def get_uid(self, data):
-        node = self.cls.nodes.get_or_none(**data)
+        node = self.cls.nodes.first_or_none(**data)
         if node:
             return node.uid
         else:
@@ -37,7 +37,7 @@ class NodeManager:
     
     def update_node_properties(self, node_id, **kwargs):
         try:
-            node = self.cls.nodes.get_or_none(uid=node_id)
+            node = self.cls.nodes.first_or_none(uid=node_id)
             if node:
                 for key, value in kwargs.items():
                     setattr(node, key, value)
@@ -50,7 +50,7 @@ class NodeManager:
     
 
     def delete_node(self, node_id):
-        node = self.cls.nodes.get_or_none(uid=node_id)
+        node = self.cls.nodes.first_or_none(uid=node_id)
         if node:
             for rel in node.relationships.all():
                 rel.delete()
