--- server/db_conn/neo4j/models/manager/relationship.py
@@ -1,13 +1,0 @@
-
-from neomodel import StructuredRel,UniqueIdProperty, Relationship,StringProperty
-
-TO = True
-FROM = False
-
-class Posting(StructuredRel):
-    uid = UniqueIdProperty()
-
-class Register(StructuredRel):
-    uid = UniqueIdProperty()
-
-
