+++ server/osint/osint.Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.8-slim
+FROM python:3.9-slim
 
 RUN apt-get update && apt-get install -y git
 COPY ./requirements.txt requirements.txt
