+++ server/src/resource/graph.py
@@ -5,12 +5,15 @@ from db_conn.neo4j.models import *
 
 bp = Blueprint('relation_graph', __name__, url_prefix='/graph')
 
-@bp.route("/node", methods=["GET"])
-def get_neo4j_data():
-    query = """
+@bp.route("/node/<string:case_id>", methods=["GET"])
+def get_neo4j_data(case_id):
+    if not case_id:
+        jsonify({'Error':'case_id did not exist'}), 404
+    query = f"""
     MATCH (n)
+    WHERE n.case_id = '{case_id}'
     OPTIONAL MATCH (n)-[r]-(m)
-    RETURN n, labels(n),TYPE(r), PROPERTIES(r), m, r.uid, n.uid, m.uid
+    RETURN n, labels(n), TYPE(r), PROPERTIES(r), m, r.uid, n.uid, m.uid
     """
     results, _ = db.cypher_query(query)
